[NFC] Make raw value struct synthesizers methods of ClangImporter::Implementation
